--- React/wbooks/scss/application.scss
@@ -1,6 +1,0 @@
-@import "equalizer";
-@import "sizes";
-@import "colors";
-@import "styles";
-@import "book";
-@import "comments";
