--- React/wbooks/scss/book.scss
@@ -1,155 +1,0 @@
-.book-default-background {
-  align-items: center;
-  background-color: $grey;
-  display: flex;
-  justify-content: center;
-}
-
-.book-img {
-  height: $book-cover-height;
-  margin-bottom: 11px;
-}
-
-.book-info {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 65px;
-  margin-right: 70px;
-  overflow-wrap: break-word;
-  width: $book-cover-width;
-
-  .book-img {
-    height: $book-cover-height;
-    margin-bottom: 11px;
-  }
-
-  .book-default-background {
-    align-items: center;
-    background-color: $grey;
-    display: flex;
-    justify-content: center;
-  }
-
-  .book-title {
-    margin-bottom: 4px;
-  }
-
-  .book-author {
-    font-size: $font-size-small;
-  }
-}
-
-.book-container {
-  display: flex;
-  flex-flow: row wrap;
-}
-
-.book-input {
-  background: $light-grey;
-  color: $grey-font;
-  font-size: $font-size-medium;
-  font-style: italic;
-  height: 38px;
-  margin-bottom: 65px;
-  margin-right: 28px;
-  padding: 0 12px;
-  width: 238px;
-}
-
-.book-search {
-  position: relative;
-
-  &::after {
-    background-image: url("./assets/search.svg");
-    background-repeat: no-repeat;
-    content: "";
-    height: 28px;
-    top: 0;
-    left: 0;
-    position: absolute;
-    width: 28px;
-  }
-}
-
-.book-detail-container {
-  display: flex;
-  margin-bottom: 36px;
-  overflow-wrap: break-word;
-
-  .book-detail-img {
-    height: 272px;
-    margin-right: 21px;
-    width: 400px;
-  }
-
-  .book-default-background {
-    align-items: center;
-    background-color: $grey;
-    display: flex;
-    justify-content: center;
-    width: 400px;
-  }
-
-  .book-detail-info-container {
-    display: flex;
-    flex-direction: column;
-
-    .book-detail-title {
-      margin-bottom: 10px;
-    }
-
-    .book-detail-author {
-      font-size: $font-size-medium-big;
-      margin-bottom: 22px;
-    }
-
-    .book-detail-year {
-      font-size: $font-size-medium-big;
-      margin-bottom: 8px;
-    }
-
-    .book-detail-genre {
-      font-size: $font-size-medium-big;
-      margin-bottom: 8px;
-    }
-
-    .book-detail-year {
-      font-size: $font-size-medium-big;
-      margin-bottom: 13px;
-    }
-
-    .book-detail-desc {
-      color: $grey-font;
-      font-size: $font-size-medium;
-      margin-bottom: auto;
-    }
-
-    .book-detail-rent-button {
-      background-color: $blue;
-      color: $white;
-      font-size: $font-size-big;
-      font-weight: bold;
-      margin-top: auto;
-      height: 38px;
-      width: 192px;
-    }
-  }
-}
-
-.green-title {
-  color: $green;
-  margin-bottom: 16px;
-}
-
-.suggestion-container {
-  display: flex;
-  flex-flow: row wrap;
-  margin-bottom: 31px;
-  padding: 0 40px;
-}
-
-.suggestion-img {
-  height: 124px;
-  margin-right: 54px;
-  width: 87px;
-}
