--- React/wbooks/scss/colors.scss
@@ -1,9 +1,0 @@
-$blue: #00adee;
-
-$grey: #d8d8d8;
-$light-grey: #f4f4f4;
-$grey-font: #9b9b9b;
-
-$green: #a5cd39;
-
-$white: #ffffff;
