--- React/wbooks/scss/sizes.scss
@@ -1,7 +1,0 @@
-$book-cover-width: 128px;
-$book-cover-height: 182px;
-
-$font-size-big: 20px;
-$font-size-medium-big: 16px;
-$font-size-medium: 14px;
-$font-size-small: 10px;
