--- React/wbooks/scss/styles.scss
@@ -1,29 +1,0 @@
-* {
-  font-family: "Roboto", sans-serif;
-}
-
-h3 {
-  font-size: $font-size-big;
-  font-weight: bold;
-}
-
-h4 {
-  font-size: $font-size-medium;
-  font-weight: bold;
-}
-
-.body-container {
-  padding: 145px 358px;
-}
-
-hr {
-  border: 1px solid $blue;
-  height: 1px;
-  margin-bottom: 31px;
-}
-
-.back-button {
-  left: 50px;
-  position: absolute;
-  top: 100px;
-}
