--- React/wbooks/src/Home/Components/BookContainer/Components/BookInfo/style.scss
@@ -1,42 +1,0 @@
-@import "/src/Components/app.scss";
-
-.book-default-background {
-  align-items: center;
-  background-color: $grey;
-  display: flex;
-  justify-content: center;
-}
-
-.book-img {
-  height: $book-cover-height;
-  margin-bottom: 11px;
-}
-
-.book-info {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 65px;
-  margin-right: 70px;
-  overflow-wrap: break-word;
-  width: $book-cover-width;
-
-  .book-img {
-    height: $book-cover-height;
-    margin-bottom: 11px;
-  }
-
-  .book-default-background {
-    align-items: center;
-    background-color: $grey;
-    display: flex;
-    justify-content: center;
-  }
-
-  .book-title {
-    margin-bottom: 4px;
-  }
-
-  .book-author {
-    font-size: $font-size-small;
-  }
-}
