+++ React/wbooks/src/WBooks.js
@@ -3,13 +3,13 @@ import { Switch, Route } from 'react-router-dom';
 import './application.css';
 
 import HomeContainer from './Home/';
-import BookDetail from './BookDetail/layout';
+import BookDetail from './BookDetail';
 
 function WBooks() {
   return (
     <Switch>
       <Route exact path="/" component={HomeContainer} />
-      <Route path="/book/:number" component={BookDetail} />
+      <Route path="/book/:id" component={BookDetail} />
     </Switch>
   );
 }
